// src/platform/plugins/shared/data_view_management/public/components/edit_index_pattern/indexed_fields_table/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the "Elastic License
 * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
 * Public License v 1"; you may not use this file except in compliance with, at
 * your election, the "Elastic License 2.0", the "GNU Affero General Public
 * License v3.0 only", or the "Server Side Public License, v 1".
 */

import { type DataViewFieldBase } from '@kbn/es-query';
import { type DataViewField } from '@kbn/data-views-plugin/public';

type IndexedFieldItemBase = Partial<DataViewField> & DataViewFieldBase;

export interface IndexedFieldItem extends IndexedFieldItemBase {
  info: string[];
  excluded: boolean;
  kbnType: string;
  isMapped: boolean;
  isUserEditable: boolean;
  hasRuntime: boolean;
}